Handle circular dependencies
To handle tthe circular dependencies the resolve method should set the
complete property of the root object to true  before trying to resolve
its dependencies. This way, after recursively calling the resolve
method for the dependencies of the root object, the ones that depend on
the root object won't try to resolve it because it will be marked as
compcomplete.