Remove new from the tags
To support the new in the tags, the resolve logic must be more complex.
Each dependency which has dependency marked as new, should be marked as
not complete because such dependencies will never be complete. This can
lead to circular dependencies which can't be resolved. To handle those
cases, the resolve method should some logic to keep track of each
hierarchy of dependencies. This will slow down the resolve method.